LearningJournal/requirements.txt
appnope==0.1.0
argh==0.26.1
beautifulsoup4==4.4.1
Chameleon==2.24
colorama==0.3.7
DateTime==4.0.1
decorator==4.0.9
docopt==0.6.2
e==1.4.5
gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
Jinja2==2.8
Mako==1.0.4
Markdown==2.6.5
MarkupSafe==0.23
passlib==1.6.5
PasteDeploy==1.5.2
path.py==8.1.2
pathtools==0.1.2
pexpect==4.0.1
pickleshare==0.6
pluggy==0.3.1
postgres==2.2.1
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pyramid==1.6.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-ipython==0.1
pyramid-jinja2==2.6.2
pyramid-mako==1.0.2
pyramid-tm==0.12.1
pytz==2016.1
pywatch==0.4
PyYAML==3.11
repoze.lru==0.6
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.12
traitlets==4.2.1
transaction==1.4.4
translationstring==1.3
venusian==1.0
virtualenv==15.0.0
waitress==0.8.10
watchdog==0.8.3
WebOb==1.5.1
WTForms==2.1
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6